+++ Classes/00-Base/EXEPackageDefinition.ps1
@@ -1,0 +1,14 @@
+class EXEPackageDefinition : Package {
+    [string]$InstallCommand
+    [string]$UninstallCommand
+    [string]$QuietUninstallCommand
+    [string]$RepairCommand
+    [int[]]$SuccessExitCodes
+    [int[]]$RebootExitCodes
+
+    EXEPackageDefinition() {
+        $this.InstallerType = 'EXE'
+        $this.SuccessExitCodes = @(0)
+        $this.RebootExitCodes = @(1641, 3010)
+    }
+}